new terminado

+++ laravel/routes/admin.php
@@ -19,7 +19,11 @@ use Inertia\Inertia;
 |
 */
 
-// Grupo que añade por defecto el prefijo de /admin/ y exige que el usuario esté registrado y sea admin
+
+//!!!!!!!!!!!!!!!!!!!Es importante que se use el middleware HasAdmin con los :!!!!!!!!!!!!!!!!!!
+
+
+// Grupo que añade por defecto el prefijo de /admin/ y exige que el usuario esté registrado
 //Route::prefix('admin')->middleware(['auth', 'verified', 'HasAdmin'])->group(function () {
 Route::prefix('admin')->group(function () {
     $indexAdminController = [AdminController::class, 'indexPaperbase'];
@@ -32,9 +36,18 @@ Route::prefix('admin')->group(function () {
     // Admin HomeAlternative princiapl
     Route::get('/home', $indexAdminController)->name("admin_homes");
 
-    // Admin Usuarios
-    Route::get('/users', $indexAdminController)->name("admin_users");
-    Route::delete('/user_delete/{id}', [AdminController::class, "deleteUser"]);
-    Route::post('/user_add', [AdminController::class, "addUser"]);
-    Route::put('/user_update/{id}', [AdminController::class, "updateUser"]);
+    //USUARIOS
+    Route::get('/users', $indexAdminController)->name("admin_users")->middleware([ 'HasAdmin:User']);
+    Route::delete('/user_delete/{id}', [AdminController::class, "deleteUser"])->middleware([ 'HasAdmin:User']);
+    Route::post('/user_add', [AdminController::class, "addUser"])->middleware([ 'HasAdmin:User']);
+    Route::put('/user_update/{id}', [AdminController::class, "updateUser"])->middleware([ 'HasAdmin:User']);
+
+    //NOTICIAS
+    Route::get('/new', $indexAdminController)->name("admin_new")->middleware(['HasAdmin:New']);
+    Route::delete('/new_delete/{id}', [AdminController::class, "deleteNew"])->middleware(['HasAdmin:New']);
+    Route::delete('/new_true_delete/{id}', [AdminController::class, "trueDeleteNew"])->middleware(['HasAdmin:New']);
+    Route::post('/new_add', [AdminController::class, "addNew"])->middleware(['HasAdmin:New']);
+    Route::put('/new_update/{id}', [AdminController::class, "updateNew"])->middleware(['HasAdmin:New']);
+    Route::post('/new_update/{id}', [AdminController::class, "updateNew"])->middleware(['HasAdmin:New']);
+    Route::post('/new_trashed', [AdminController::class, "toggleTrashedNew"])->middleware(['HasAdmin:New']);
 });
